refactor the module conf structure
define conf blocks (secure_token_akamai / secure_token_cloudfront) that
create nginx variables for the tokens.
the tokens are then selected on a specific location by the use of the
secure_token directive.
this change enables the use of multiple token configurations on a single
location, via the use of nginx map directive.