missing tokens bug fix
if an input buffers ends on the scheme of the url, last_url_char stays
'\n' from the previous iteration. this makes the m3u8 processor go back
to the initial state. on the next iteration, it will start the url
parsing state machine again, while it's in the middle of the scheme. the
url parser will therefore see only part of the scheme (e.g. tps instead
of https). it will enter the 'non http' state, and skip adding the
token.